# fragtale-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fragtale_api",
    "description": "Fragtale API",
    "license": {
      "name": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/topics/{topic_id}/events": {
      "get": {
        "tags": [
          "web_socket"
        ],
        "summary": "Open a WebSocket connection for publishing events.",
        "description": "Publisher identifier is derived from authentication.",
        "operationId": "publish_event_to_topic",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols to websocket."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "http"
        ],
        "summary": "Publish event document.",
        "description": "Please note the `correlation-token` and `location` header if you need to\nfind an event from a different topic that is the result of processing this event.\n\nPublisher identifier is derived from authentication.",
        "operationId": "publish_event_to_topic",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Importance of the published event. 0-100 where 100 is most important.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "Event Descriptor SemVer the event is expected to comply to. (E.g. major.minor).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target",
            "in": "query",
            "description": "Expected target topic of correlated event processing.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok. Successfully published event and returning response of correlated request.",
            "content": {
              "application/json": {}
            }
          },
          "204": {
            "description": "No content. Successfully published event.",
            "headers": {
              "correlation-token": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque token that can be used to correlate events."
              }
            }
          },
          "303": {
            "description": "See other. Correlated result did not appear before the timeout. Poll linked resource to keep trying.",
            "links": {
              "Location": {
                "operation_id": "by_topic_and_correlation_token",
                "parameters": {
                  "correlation_token": "string",
                  "topic_id": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          },
          "403": {
            "description": "Forbidden: Authorization failure."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/topics/{topic_id}/subscribe": {
      "get": {
        "tags": [
          "web_socket"
        ],
        "summary": "Open a WebSocket connection for subscribing to new events.",
        "description": "Consumer identifier is derived from authentication.",
        "operationId": "subscribe_to_topic",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Only consider events newer than this in epoch milliseconds.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "Event Descriptor SemVer that the client prefers (major.minor).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols to websocket."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized: Authentication failure."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}